fixed cypress tests running, pass
