+++ src/Components/Attendees.js
@@ -1,3 +1,11 @@
-export default function Attendees() {
-  return;
+import Attendee from "./Attendee"
+
+export default function Attendees({ event, updateEventAttendance, attendees }) {
+  return (
+    <div className="attendees">
+      {attendees.map((attendee, index) => (
+        <Attendee updateEventAttendance={updateEventAttendance} attendee={attendee} event={event} key={index} />
+      ))}
+    </div>
+  )
 }
